requirements.txt
git+https://github.com/ellisonleao/pyshorteners.git
PyYAML

